storage: move tscache-related methods to new file
Updating and applying the timestamp cache are two sides of the same
coin. It makes sense that these methods should live right next to
each other.

Release note: None
